# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "schedule-maker-api",
        "contact": {
            "name": "Sanzhar",
            "url": "https://t.me/svnzh1k"
        },
        "version": "1.0"
    },
    "host": "localhost:7070",
    "paths": {
        "/lectures/{subjectId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "Получить список лекций",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subject ID",
                        "name": "subjectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Lecture"
                            }
                        }
                    }
                }
            }
        },
        "/practices/{lectureId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Practices"
                ],
                "summary": "Получить список практисов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Practice"
                            }
                        }
                    }
                }
            }
        },
        "/subjects": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjects"
                ],
                "summary": "Получить список предметов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Subject"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Lecture": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lecturer": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "main.Practice": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lecture_id": {
                    "type": "integer"
                },
                "practice_teacher": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "main.Subject": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "subject_name": {
                    "type": "string"
                }
            }
        }
    }
}